move things out of extradencommands
